set up html page for the job

+++ src/jsonPrser.ts
@@ -1,10 +1,12 @@
-import { subventions } from './subventions';
 import fs from "fs";
 import path from "path";
 import { stringify } from "csv-stringify";
 import { Logger } from "./logger";
+import archiver from "archiver";
 type AnyDict = { [key: string]: any };
 let arraysStorage: AnyDict
+const logger = new Logger("aggregateParser");
+
 const deepTraverse = (obj: AnyDict, arrayKey?: string) => {
 
   const objEntries = Object.entries(obj);
@@ -12,41 +14,96 @@ const deepTraverse = (obj: AnyDict, arrayKey?: string) => {
   for (let [key, value] of objEntries) {
     if (!!value) {
       if (Array.isArray(value)) {
-        arraysStorage?.[key] || (arraysStorage[key] = []);
+        if (typeof key === "string") {
+          arraysStorage?.[key] || (arraysStorage[key] = []);
+        }
         value.forEach(element => {
-          deepTraverse(value, key)
+          deepTraverse(element, key)
         });
-      }
+        const newValue = value.map((item) => {
+          if (typeof item === "object") {
+            const ind = arraysStorage[key].push({ index: arraysStorage[key].length + 1, ...item });
+            return ind;
+          } else {
+            return item;
+          }
+        });
+        obj[key] = newValue.join("|");
 
-      if (typeof value === "object") {
+      } else if (typeof value === "object") {
         if (typeof key === "string") {
           arraysStorage?.[key] || (arraysStorage[key] = []);         
         }
         deepTraverse(value);
         const keyToPusInto = arrayKey || key;
 
-        arraysStorage[keyToPusInto].push(value);
+        const ind = arraysStorage[keyToPusInto].push({ index: arraysStorage[keyToPusInto].length + 1, ...value });
+        obj[key] = ind
       }
     }
 
   }
 }
+const writeCSV = (data: AnyDict[], filePath: string): Promise<void> => {
+  // Replace newline characters in each string field of each object.
+  const sanitizedData = data.map(obj => {
+    const sanitizedObj: AnyDict = {};
+    for (const [key, value] of Object.entries(obj)) {
+      if (typeof value === 'string') {
+        sanitizedObj[key] = value.replace(/\n/g, ' ');  // Replace newline with space
+      } else {
+        sanitizedObj[key] = value;
+      }
+    }
+    return sanitizedObj;
+  });
 
+  return new Promise<void>((resolve, reject) => {
+    stringify(sanitizedData, { header: true }, (err, output) => {
+      if (err) {
+        reject(err);
+      } else {
+        fs.writeFileSync(filePath, output);
+        resolve();
+      }
+    });
+  });
+};
+
+
+const createZip = (sourceDir: string, outputFilePath: string): Promise<void> => {
+  return new Promise<void>((resolve, reject) => {
+    const output = fs.createWriteStream(outputFilePath);
+    const archive = archiver('zip', {
+      zlib: { level: 9 }, // Compression level
+    });
+
+    output.on('close', () => {
+      resolve();
+    });
+
+    archive.on('error', (err) => {
+      reject(err);
+    });
+
+    archive.pipe(output);
+    archive.directory(sourceDir, false);
+    archive.finalize();
+  });
+};
 
 
 export async function aggregateParser(jsonFilePath: string): Promise<void> {
-  const logger = new Logger("aggregateParser");
   try {
     arraysStorage = {
       subventions: [],
     };
     const outputDirectory = path.join(__dirname, "outputs/explodedData");
-    const fs = require("fs");
-
-
-    fs.readdirSync(outputDirectory).forEach((file: any) => {
-      fs.unlinkSync(path.join(outputDirectory, file));
-    });
+    if (fs.existsSync(outputDirectory)) {
+      fs.readdirSync(outputDirectory).forEach((file: any) => {
+        fs.unlinkSync(path.join(outputDirectory, file));
+      });
+    }
 
     fs.mkdirSync(outputDirectory, { recursive: true });
 
@@ -58,15 +115,32 @@ export async function aggregateParser(jsonFilePath: string): Promise<void> {
       arraysStorage.subventions.push(item);
       deepTraverse(item);
     }
+    const date = new Date();
+    const timeStamp = `${date.getDate()}-${date.getMonth()}-${date.getFullYear()}`;
 
     //write one JSON file per key of arraysStorage in the output directory
     for (let key in arraysStorage) {
-      const filePath = path.join(outputDirectory, `${key}.json`);
-      logger.info(`Writing ${filePath}`);
+      const jsonFilePath = path.join(outputDirectory, `${timeStamp}-${key}.json`);
+      const csvFilePath = path.join(outputDirectory, `${timeStamp}-${key}.csv`);
 
-      fs.writeFileSync(filePath, JSON.stringify(arraysStorage[key], null, 2));
+      // Remove empty arrays
+      Object.keys(arraysStorage).forEach((key) => {
+        if (arraysStorage[key].length === 0) {
+          delete arraysStorage[key];
+        }
+      });
+
+      logger.info(`Writing ${jsonFilePath}`);
+      fs.writeFileSync(jsonFilePath, JSON.stringify(arraysStorage[key], null, 2));
+
+      logger.info(`Writing ${csvFilePath}`);
+      await writeCSV(arraysStorage[key], csvFilePath);
     }
+    const zipFilePath = path.join(__dirname, `outputs/explodedData/${timeStamp}-explodedData.zip`);
 
+    logger.info(`Creating ZIP archive ${zipFilePath}`);
+
+    await createZip(outputDirectory, zipFilePath);
 
     logger.verbose("Done");
   } catch (error) {
